// photovoltaicV2/src/app/services/models/solar-panel.ts
/* tslint:disable */
/* eslint-disable */
export interface SolarPanel {
  company: {id : string};
  height: number;
  id: number;
  image: string;
  maximum_current: number;
  maximum_voltage: number;
  model: string;
  nominal_power: number;
  price: number;
  quantity: number;
  type_cell: string;
  weight: number;
  width: number;
}